requirements.txt
asgiref==3.4.1
Django==3.2.7
django-autoslug==1.9.8
django-debug-toolbar==3.2.2
django-localflavor==3.1
django-shopping-cart==0.1
Pillow==8.3.2
python-stdnum==1.16
pytz==2021.1
sqlparse==0.4.2
